separated projects component
