+++ src/index.html
@@ -7,6 +7,7 @@
     <title>Ashley's Personal Bio Site</title>
     <link href="https://fonts.googleapis.com/css?family=Abel|Gayathri&display=swap" rel="stylesheet">
 </head>
+
 <body>
     <header>
         <h1 id="homeDefault">Ashley Claiborne's Personal Site</h1>
@@ -19,7 +20,6 @@
         </ul>
     </nav>
 
-    <!-- Primary View -->
     <div id="mainDiv" class="pb-4">
         <div class="fullPage navToBio" id="bioPage">
             <h3>Career Bio</h3>
@@ -67,9 +67,7 @@
 
         </div>
     </div>
-    <!-- End Primary View -->
 
-    <!-- Footer -->
     <footer class="page-footer fixed-bottom">
         <h5>Contact</h5>
         <div class="footerLinks">
@@ -79,6 +77,5 @@
             <a href="https://codepen.io/aclai4067" title="codepen"><i class="fab fa-codepen"></i></a>
         </div>
     </footer>
-    <!-- End Footer -->
 </body>
 </html>
