+++ src/javascripts/main.js
@@ -5,40 +5,9 @@ import '../styles/main.scss';
 import $ from 'jquery';
 import firebase from 'firebase';
 import apiKeys from './helpers/apiKeys.json';
-import projectData from './helpers/data/projectData';
+import projects from './components/projects/projects';
 
 const bodyDivs = document.getElementsByClassName('fullPage');
-
-const printToDom = (divId, textToPrint) => {
-  const selectedDiv = document.getElementById(divId);
-  selectedDiv.innerHTML = textToPrint;
-};
-
-const createProjectCards = () => {
-  let cardsToPrint = '';
-  projectData.getProjects().then((projectArr) => {
-    console.log(projectArr);
-    projectArr.forEach((projectObj) => {
-      if (projectObj.available) {
-        cardsToPrint += `
-          <div class="card col-sm-5 m-sm-4 m-1 p-0 projectCard ${projectObj.available}">
-            <img class="card-img-top p-2" src=${projectObj.screenshot} alt="${projectObj.title} screenshot" />
-            <div class="card-body">
-                <h4 class="card-title"> ${projectObj.title} </h4>
-                <p>${projectObj.description}</p>
-                <p>${projectObj.technologiesUsed}</p>
-            </div>
-            <footer class="card-footer projectLinks">
-                <a href=${projectObj.url}>URL</a>
-                <a href=${projectObj.githubUrl}>GitHub Repo</a>
-            </footer>
-          </div>
-        `;
-      }
-    });
-    printToDom('projectsBody', cardsToPrint);
-  }).catch((err) => console.error(err));
-};
 
 const toggle = (e) => {
   const selectedLink = $(e.target).attr('id');
@@ -54,7 +23,7 @@ const toggle = (e) => {
 
 const init = () => {
   firebase.initializeApp(apiKeys.firebaseConfig);
-  createProjectCards();
+  projects.createProjectCards();
   $('body').on('click', '.navLinks', toggle);
 };
